Refactor bounds into own module
